// src/types/appearance.ts
export interface Colors {
  light: {
    primary: string;
    secondary: string;
    background: string;
    foreground: string;
    card: string;
    cardForeground: string;
    muted: string;
    mutedForeground: string;
    popover: string;
    popoverForeground: string;
    border: string;
    input: string;
    ring: string;
  };
  dark: {
    primary: string;
    secondary: string;
    background: string;
    foreground: string;
    card: string;
    cardForeground: string;
    muted: string;
    mutedForeground: string;
    popover: string;
    popoverForeground: string;
    border: string;
    input: string;
    ring: string;
  };
}